requirements.txt
streamlit
tensorflow
opencv-python
pandas
numpy
scipy
scikit-learn
matplotlib
seaborn
pillow